🎨 (feature-api) Moved all user related api calls to auth route
