+++ backend/api/apps/src/index.js
@@ -7,116 +7,12 @@ const app = express();
 app.use(bodyParser.json());
 app.use(Cors())
 
+//Connect to db here
+
 const authRoute = require("./routes/auth")
 
 //router middleware
 app.use("/api/user",authRoute);
-
-/**
- * Add a user
- * Adds a user with an Name, Age
- */
- app.post('/user/signup',async(req,res)=>{
-    fs.readFile( __dirname + "/users.json", 'utf8', function (err, data) {
-        if(err) {
-            return console.log(err);
-        }
-        let d = JSON.parse(data);
-        let last = Object.keys(d)[Object.keys(d).length-1].slice(-1)
-        let userid = "user"+(parseInt(last)+1);
-        let user = {
-            [userid]:{
-                "name": req.body.name,
-                "age": req.body.age
-            }
-        }
-        d[userid] = user[userid];
-        fs.writeFile(__dirname + "/users.json", JSON.stringify(d,null,2), function(erra) {
-            if(erra) {
-                return console.log(erra);
-            }
-        });
-
-        return res.status(200).end(JSON.stringify(d,null,2));
-    });
-})
-
-/**
- * Log a user in
- * Logs the user in with their password and username
- */
-app.post("/user/login",async(req,res)=>{
-    fs.readFile( __dirname + "/users.json", 'utf8', function (err, data) {
-        if(err) {
-            return console.log(err);
-        }
-        let d = JSON.parse(data);
-        for(var i = 0;i < Object.keys(d).length;i++){
-            if(d[Object.keys(d)[i]].name === req.body.name){
-                return res.status(200).end(JSON.stringify("User logged in successfully"));
-            }
-        }
-
-        return res.status(400).end(JSON.stringify("Login Failed",null,2));
-    });
-})
-
-/**
- * Delete an item
- * Uses the user id delete the user
- */
- app.post('/user/delete',async(req,res)=>{
-    fs.readFile( __dirname + "/users.json", 'utf8', function (err, data) {
-        if(err) {
-            return console.log(err);
-        }
-        let d = JSON.parse(data);
-        for(var i = 0;i < Object.keys(d).length;i++){
-            if(Object.keys(d)[i] == req.body.userid){
-                delete d[req.body.userid];
-                break;
-            }
-        }
-        fs.writeFile(__dirname + "/users.json", JSON.stringify(d,null,2), function(erra) {
-            if(erra) {
-                return console.log(erra);
-            }
-        }); 
-
-        return res.status(200).end(JSON.stringify(d,null,2));
-    });
-})
-
-/**
- * Update a user
- * Uses the user id to update the user
- */
- app.post('/user/update',async(req,res)=>{
-    fs.readFile( __dirname + "/users.json", 'utf8', function (err, data) {
-        if(err) {
-            return console.log(err);
-        }
-        let d = JSON.parse(data);
-        for(var i = 0;i < Object.keys(d).length;i++){
-            if(Object.keys(d)[i] == req.body.userid){
-                if(req.body.name != undefined){
-                    d[req.body.userid].name = req.body.name;
-                }
-                if(req.body.age != undefined){
-                    d[req.body.userid].age = req.body.age;
-                }
-                break;
-            }
-        }
-        fs.writeFile(__dirname + "/users.json", JSON.stringify(d,null,2), function(erra) {
-            if(erra) {
-                return console.log(erra);
-            }
-        }); 
-
-        return res.status(200).end(JSON.stringify(d,null,2));
-    });
-})
 
 /**
  * Get all items for a user
